requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
gptcache==0.1.17
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
langchain==0.0.142
llama-index==0.5.21
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.0
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
slack-bolt==1.16.4
slack-sdk==3.20.2
SQLAlchemy==1.4.46
sqlparse==0.4.3
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
